igor export & postprocessing to its own class
