+++ tb_mean_field_hubbard/__init__.py
@@ -1,3 +1,4 @@
 from .mfh import MeanFieldHubbardModel
+from .utils import create_itx
 
 __version__ = "1.5.1"
